requirements.txt
# After installing dependencies, run: python -m playwright install
playwright>=1.45.0
pytest>=8.3.4
pydantic>=2.8.2
pydantic-core>=2.20.1
pyyaml>=6.0.1
python-dotenv>=1.0.1
sqlalchemy>=2.0.31
apscheduler>=3.10.4
requests>=2.32.3
tenacity>=8.3.0
beautifulsoup4>=4.12.3
fastapi>=0.111.0
itsdangerous>=2.2.0
uvicorn[standard]>=0.30.1
jinja2>=3.1.4
python-multipart>=0.0.9
openpyxl>=3.1.5
httpx>=0.27.2
playwright-stealth>=2.0.0
